* Refactor components * Add eslint and .eslintrc * Fix code style issues * WIP logo and design
