+++ src/components/logo/logo.scss
@@ -1,0 +1,3 @@
+.logo {
+  max-width: 300px;
+}
