+++ src/components/search/Search.scss
@@ -7,29 +7,18 @@ $accent-color: #4083c7;
     box-sizing: border-box;
 }
 
-html, body {
-    margin: 0;
-    padding: 0;
-}
-
 h1.title {
     color: $accent-color;
     margin-top: 30px;
     margin-bottom: 60px;
     padding: 0 60px
 }
+
 #main {
     text-align: center;
     margin: 0 auto;
     max-width: 720px;
     padding: 20px 25px 100px;
-}
-.logo {
-    position: absolute;
-    max-width: 100px;
-    margin-top: 10px;
-    top: 0;
-    left: 0;
 }
 
 .search-form {
@@ -97,11 +86,11 @@ h1.title {
         border-bottom: 1px dashed #666;
         text-align: left;
         padding-bottom: 30px;
-        
+
         &:not(:first-child) {
             margin: 30px 0;
         }
-    
+
         ul {
             list-style: none;
         }
